Restructure GAE deployment and Datastore interface
Description:
This is a _very_ big check-in. :grimacing:

Deployments have been changed so that the Google App Engine
standard environment is the target (instead of flexible). This
allows for the free tier to be utilized.

Additionally, major refactoring work has been done on database
methods with the switch over to Datastore. The emulator tests
are still non-functional and have been temporarily commented out
in TravisCI.

Coverage results are not being sent to Coveralls at the moment
due to that service being down for maintenance.

Note: Initial FAQ handler included
Note: Test coverage decreased significantly
Note: Alternative deployments/structures are under consideration